refactor: auth_service - auth
