+++ internal-services/account/auth/auth.go
@@ -1,4 +1,4 @@
-package middleware
+package auth
 
 import (
 	"bytes"
